Refactoring common encryption utility functions out of encryption_internal (#70)
