+++ cpp/src/parquet/encryption/crypto_factory.cc
@@ -21,7 +21,7 @@
 #include "arrow/util/string.h"
 
 #include "parquet/encryption/crypto_factory.h"
-#include "parquet/encryption/encryption_internal.h"
+#include "parquet/encryption/encryption_utils.h"
 #include "parquet/encryption/file_key_unwrapper.h"
 #include "parquet/encryption/file_system_key_material_store.h"
 #include "parquet/encryption/key_toolkit_internal.h"
