+++ cpp/src/parquet/encryption/internal_file_decryptor.cc
@@ -20,6 +20,7 @@
 #include "arrow/util/logging.h"
 #include "parquet/encryption/encryption.h"
 #include "parquet/encryption/encryption_internal.h"
+#include "parquet/encryption/encryption_utils.h"
 #include "parquet/metadata.h"
 
 namespace parquet {
