+++ cpp/src/parquet/encryption/internal_file_encryptor.cc
@@ -18,6 +18,7 @@
 #include "parquet/encryption/internal_file_encryptor.h"
 #include "parquet/encryption/encryption.h"
 #include "parquet/encryption/encryption_internal.h"
+#include "parquet/encryption/encryption_utils.h"
 
 namespace parquet {
