+++ cpp/src/parquet/encryption/internal_file_encryptor.h
@@ -22,6 +22,7 @@
 #include <string>
 
 #include "parquet/encryption/encryption.h"
+#include "parquet/encryption/encryption_internal.h"
 #include "parquet/metadata.h"
 
 namespace parquet {
